Modular version of dash app
