+++ ai/monitoring/app_scripts/constants.py
@@ -1,0 +1,14 @@
+# Custom color palette
+colors = {
+    'background': '#222',
+    'text': '#fff',
+    'primary': '#636EFA',
+    'secondary': '#EF553B',
+    'tertiary': '#00CC96',
+    'quaternary': '#AB63FA',
+    'warning': '#FFA15A'
+}
+
+# Interval settings (in milliseconds)
+LIVE_UPDATE_INTERVAL = 10 * 1000  # 10 seconds
+DATA_REFRESH_INTERVAL = 60 * 1000  # 1 minute
